requirements.txt
humanize
discord
pillow
requests
python-dotenv